fix(oci): auth client creation (#158)
## Description

The use of the DefaultClient is resulting in a shared reference between
multiple remotes. In the case of publishing - the CredentialStore is
always modified by the last remote created - which can result in
authentication for whichever remote was created first being invalid.

## Related Issue

Relates to https://github.com/zarf-dev/zarf/issues/3647