fix: allow `o.PullPath` to pull paths with sub directories (#175)
## Description

There is currently an issue in `o.pullPath` that forces users to create
the entire path ahead of time before pulling an image. This can be
problematic for layers that define a multi directory path such as
`components/baseline.tar`. The caller would have to know to create the
`components` directory ahead of time.

Signed-off-by: Austin Abro <AustinAbro321@gmail.com>